Move word package from content to screens
