+++ app/src/main/kotlin/com/jehutyno/yomikata/screens/word/WordContract.kt
@@ -1,4 +1,4 @@
-package com.jehutyno.yomikata.screens.content.word
+package com.jehutyno.yomikata.screens.word
 
 import androidx.lifecycle.LiveData
 import com.jehutyno.yomikata.BasePresenter
